Move bank extractors and update skill
