+++ lib/src/bank_support/extractors/extract_bob.dart
@@ -1,4 +1,4 @@
-import '../shared/models/spending_model.dart';
+import '../../shared/models/spending_model.dart';
 
 Iterable<Transaction> extractBOBMessages(
   Iterable<String> bobMessages,
