+++ test/extract_test.dart
@@ -1,8 +1,8 @@
 import 'package:flutter_test/flutter_test.dart';
+import 'package:messages_wallet/src/bank_support/extractors/extract_axis.dart';
+import 'package:messages_wallet/src/bank_support/extractors/extract_bob.dart';
+import 'package:messages_wallet/src/bank_support/extractors/extract_cosmos.dart';
 import 'package:messages_wallet/src/shared/models/spending_model.dart';
-import 'package:messages_wallet/src/utils/extract_axis.dart';
-import 'package:messages_wallet/src/utils/extract_bob.dart';
-import 'package:messages_wallet/src/utils/extract_cosmos.dart';
 
 import 'test_data.dart';
